refactor: extract `SeekTo` to its own file
