+++ core/src/main/scala/com/banno/kafka/consumer/ConsumerOps.scala
@@ -14,89 +14,22 @@
  * limitations under the License.
  */
 
-package com.banno.kafka.consumer
+package com.banno.kafka
+package consumer
 
-import cats._
-import cats.implicits._
-import fs2._
+import cats.*
+import cats.effect.*
+import cats.syntax.all.*
 
-import cats.effect._
+import fs2.*
 
-import scala.jdk.CollectionConverters._
-import scala.concurrent.duration._
-import org.apache.kafka.common._
+import scala.jdk.CollectionConverters.*
+import scala.concurrent.duration.*
+import org.apache.kafka.common.*
 import org.apache.kafka.common.errors.WakeupException
-import org.apache.kafka.clients.consumer._
-import com.banno.kafka._
+import org.apache.kafka.clients.consumer.*
 import fs2.concurrent.{Signal, SignallingRef}
 import org.typelevel.log4cats.slf4j.Slf4jLogger
-
-sealed trait SeekTo
-
-object SeekTo {
-
-  private case object Beginning extends SeekTo
-  private case object End extends SeekTo
-  private case class Timestamps(timestamps: Map[TopicPartition, Long], default: SeekTo)
-      extends SeekTo
-  private case class Timestamp(timestamp: Long, default: SeekTo) extends SeekTo
-  private case class Committed(default: SeekTo) extends SeekTo
-  private case class Offsets(offsets: Map[TopicPartition, Long], default: SeekTo) extends SeekTo
-
-  def beginning: SeekTo = Beginning
-  def end: SeekTo = End
-  def timestamps(timestamps: Map[TopicPartition, Long], default: SeekTo): SeekTo =
-    Timestamps(timestamps, default)
-  def timestamp(timestamp: Long, default: SeekTo): SeekTo = Timestamp(timestamp, default)
-  def committed(default: SeekTo): SeekTo = Committed(default)
-  def offsets(offsets: Map[TopicPartition, Long], default: SeekTo): SeekTo =
-    Offsets(offsets, default)
-  def timestampBeforeNow[F[_]: Clock: Functor](duration: FiniteDuration, default: SeekTo): F[SeekTo] =
-    Clock[F].realTime.map(now =>
-      timestamp(now.toMillis - duration.toMillis, default))
-
-  def seek[F[_]: Monad](
-      consumer: ConsumerApi[F, _, _],
-      partitions: Iterable[TopicPartition],
-      seekTo: SeekTo
-  ): F[Unit] =
-    seekTo match {
-      case Beginning =>
-        consumer.seekToBeginning(partitions)
-      case End =>
-        consumer.seekToEnd(partitions)
-      case Offsets(offsets, default) =>
-        partitions.toList.traverse_(
-          tp =>
-            offsets
-              .get(tp)
-              //p could be mapped to an explicit null value
-              .flatMap(Option(_))
-              .fold(SeekTo.seek(consumer, List(tp), default))(o => consumer.seek(tp, o))
-        )
-      case Timestamps(ts, default) =>
-        for {
-          offsets <- consumer.partitionQueries.offsetsForTimes(ts)
-          () <- seek(
-            consumer,
-            partitions,
-            Offsets(offsets.view.mapValues(_.offset).toMap, default)
-          )
-        } yield ()
-      case Timestamp(timestamp, default) =>
-        val timestamps = partitions.map(p => (p, timestamp)).toMap
-        seek(consumer, partitions, Timestamps(timestamps, default))
-      case Committed(default) =>
-        for {
-          committed <- consumer.partitionQueries.committed(partitions.toSet)
-          () <- seek(
-            consumer,
-            partitions,
-            Offsets(committed.view.mapValues(_.offset).toMap, default)
-          )
-        } yield ()
-    }
-}
 
 case class PartitionQueriesOps[F[_]](consumer: PartitionQueries[F]) {
